moves rac_user to users
