+++ project_electron/orgs/templates/accession/create.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 
 {% block h1_title %}{{form.accession_number.value}} <small>{{ form.extent_files.value }} files, {{ form.extent_size.value|filesizeformat }}</small>{% endblock %}
