+++ project_electron/orgs/templates/appraise/main.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 
 {% block h1_title %}Appraisal Queue <small>{{uploads|length}} transfers</small>{% endblock %}
