+++ project_electron/orgs/templates/orgs/create.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 
 {% block h1_title %}Add an Organization{% endblock %}
