+++ project_electron/orgs/templates/orgs/detail.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 {% load tz static %}
 
 {% block h1_title %}
