+++ project_electron/orgs/templates/orgs/transfers.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 {% load tz %}
 
 {% block h1_title %}
