+++ project_electron/orgs/templates/orgs/update.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 {% load static %}
 {% block h1_title %}{{ object.name }}<small>Organization Edit</small>{% endblock %}
