+++ project_electron/orgs/templates/rest_framework/api.html
@@ -1,4 +1,4 @@
-{% extends "transfer_app/base.html" %}
+{% extends "transfers/base.html" %}
 {% load static %}
 {% load i18n %}
 {% load rest_framework %}
