+++ project_electron/orgs/templates/rights/manage.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 
 {% block h1_title %}{{organization}} <small>PREMIS Rights Statement</small>{% endblock %}
