+++ project_electron/orgs/templates/transfer_app/base.html
@@ -1,16 +1,16 @@
 <!DOCTYPE html>
 <html>
   <head>
-    {% include 'transfer_app/head.html' %}
+    {% include 'transfers/head.html' %}
     {% block extra_css %}{% endblock %}
   </head>
   <body class="hold-transition skin-blue sidebar-mini">
     <div class="wrapper">
       <!-- Main Header -->
-      {% include 'transfer_app/header.html' %}
+      {% include 'transfers/header.html' %}
 
       <!-- Left Sidebar -->
-      {% include 'transfer_app/left_sidebar.html' %}
+      {% include 'transfers/left_sidebar.html' %}
 
       <!-- Content Wrapper. Contains page content -->
       <div class="content-wrapper">
@@ -29,22 +29,22 @@
         <!-- Main content -->
         <section class="content">
 
-        {% include 'transfer_app/messages.html' %}
+        {% include 'transfers/messages.html' %}
         
         {% block content %}{% endblock %}
         </section>
       </div>
 
       <!-- Main Footer -->
-      {% include 'transfer_app/footer.html' %}
+      {% include 'transfers/footer.html' %}
 
       
-      {% include 'transfer_app/sidebar.html' %}
+      {% include 'transfers/sidebar.html' %}
 
       {% block modals %}{% endblock %}
 
       <!-- REQUIRED JS SCRIPTS -->
-      {% include 'transfer_app/js_scripts.html' %}
+      {% include 'transfers/js_scripts.html' %}
       {% block extra_js %}{% endblock %}
 
     </div>
