+++ project_electron/orgs/templates/transfer_app/dashboard.html
@@ -44,7 +44,7 @@
           <div class="row">
             <div class="col-sm-3 col-xs-6">
               <div class="description-block border-right">
-                {% include 'transfer_app/trend_display.html' with trend=count_trend %}
+                {% include 'transfers/trend_display.html' with trend=count_trend %}
                 <h5 class="description-header">{{ org.upload_count_by_month|last }}</h5>
                 <span class="description-text">Transfers in {% now "F" %}</span>
               </div>
@@ -52,7 +52,7 @@
 
             <div class="col-sm-3 col-xs-6">
               <div class="description-block border-right">
-                {% include 'transfer_app/trend_display.html' with trend=size_trend %}
+                {% include 'transfers/trend_display.html' with trend=size_trend %}
                 <h5 class="description-header">{{ org.upload_size_by_month|last|floatformat:"2" }}</h5>
                 <span class="description-text">Gigabytes in {% now "F" %}</span>
               </div>
