+++ project_electron/orgs/templates/transfer_app/login_base.html
@@ -2,7 +2,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-  {% include 'transfer_app/head.html' %}
+  {% include 'transfers/head.html' %}
 
     <!-- iCheck -->
     <link rel="stylesheet" href="{% static "plugins/iCheck/square/blue.css" %}">
@@ -19,7 +19,7 @@
     <div class="login-credit">Image by Soerfm <a href="https://creativecommons.org/licenses/by-sa/3.0">CC BY-SA 3.0</a>, <a href="https://commons.wikimedia.org/wiki/File%3AAurora_borealis_above_Lyngenfjorden%2C_2012_March-3.jpg">via Wikimedia Commons</a></div>
 
     <!-- REQUIRED JS SCRIPTS -->
-    {% include 'transfer_app/js_scripts.html' %}
+    {% include 'transfers/js_scripts.html' %}
 
     <script src="{% static "plugins/iCheck/icheck.min.js" %}"></script>
     <script>
