+++ project_electron/orgs/templates/transfer_app/transfer_detail.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/base.html' %}
+{% extends 'transfers/base.html' %}
 {% load tz static %}
 {% load util %}
