+++ project_electron/orgs/templates/users/password_reset_complete.html
@@ -1,4 +1,4 @@
-{% extends 'transfer_app/login_base.html' %}
+{% extends 'transfers/login_base.html' %}
 
 {% block content %}
 <div class="login-box-body">
