+++ project_electron/orgs/transfers/cron.py
@@ -15,7 +15,7 @@ class MyCronJob(CronJobBase):
     RUN_EVERY_MINS = 1 # every 2 hours
 
     schedule = Schedule(run_every_mins=RUN_EVERY_MINS)
-    code = 'transfer_app.my_cron_job'    # a unique code
+    code = 'transfers.my_cron_job'    # a unique code
 
     def do(self):
         Pter.cron_open()
